main to emp error resolve

+++ BugTracking/src/main/java/com/cg/bugtracking/service/ProjectServiceImpl.java
@@ -17,13 +17,14 @@ import com.cg.bugtracking.entity.Admin;
 import com.cg.bugtracking.entity.Project;
 import com.cg.bugtracking.exception.NoSuchAdminFoundException;
 import com.cg.bugtracking.exception.NoSuchProjectFoundException;
+import com.cg.bugtracking.exception.NotAdminException;
 
 @Service
 public class ProjectServiceImpl implements ProjectService {
 
 	private static final Logger LOG = LogManager.getLogger(ProjectServiceImpl.class);
 	private static final String NO_PROJECT_FOUND = "Project not found.";
-	private static final String NO_ADMIN_FOUND = "Admin ID not found.";
+	private static final String NOT_ADMIN = "Admin ID not found.";
 
 	@Autowired
 	private ProjectRepository pRepo;
@@ -35,16 +36,21 @@ public class ProjectServiceImpl implements ProjectService {
 	private ModelMapper modelMapper;
 
 	@Override
-	public ProjectDTO createProject(ProjectDTO prjDTO) {
+	public ProjectDTO createProject(ProjectDTO prjDTO, long adminId) throws NotAdminException {
 		Project prj = modelMapper.map(prjDTO, Project.class);
-		LOG.info("Saving project");
-		pRepo.save(prj);
-		LOG.info("Saved Return project");
-		return prjDTO;
+		Optional<Admin> findAdmin = aRepo.findById(adminId);
+		if (findAdmin.isPresent()) {
+			LOG.info("Saving project");
+			pRepo.save(prj);
+			LOG.info("Saved Return project");
+			return prjDTO;
+		} else {
+			throw new NotAdminException(NOT_ADMIN);
+		}
 	}
 
 	@Override
-	public ProjectDTO getProjectById(long prjId, long adminId) throws NoSuchProjectFoundException, NoSuchAdminFoundException {
+	public ProjectDTO getProjectById(long prjId, long adminId) throws NoSuchProjectFoundException, NotAdminException {
 		Optional<Project> prj = pRepo.findById(prjId);
 		Optional<Admin> adm = aRepo.findById(adminId);
 		if (adm.isPresent()) {
@@ -54,46 +60,63 @@ public class ProjectServiceImpl implements ProjectService {
 			} else {
 				throw new NoSuchProjectFoundException(NO_PROJECT_FOUND);
 			}
-		}else {
-			throw new NoSuchAdminFoundException(NO_ADMIN_FOUND);
+		} else {
+			throw new NotAdminException(NOT_ADMIN);
 		}
 	}
 
 	@Override
-	public List<ProjectDTO> getAllProjects() {
-		LOG.info("Returning all projects");
-		return pRepo.findAll().stream().map(prj -> modelMapper.map(prj, ProjectDTO.class)).collect(Collectors.toList());
+	public List<ProjectDTO> getAllProjects(long adminId) throws NotAdminException {
+		if (aRepo.existsById(adminId)) {
+			LOG.info("Returning all projects");
+			return pRepo.findAll().stream().map(prj -> modelMapper.map(prj, ProjectDTO.class))
+					.collect(Collectors.toList());
+		} else {
+			throw new NotAdminException(NOT_ADMIN);
+		}
 
 	}
 
 	@Override
-	public ProjectDTO updateProject(long id, ProjectDTO pDTO) throws NoSuchProjectFoundException {
+	public ProjectDTO updateProject(long id, ProjectDTO pDTO, long adminId)
+			throws NoSuchProjectFoundException, NotAdminException {
 
 		Optional<Project> prjToUpdate = pRepo.findById(id);
 		Project prj = modelMapper.map(pDTO, Project.class);
-
-		if (prjToUpdate.isPresent()) {
-			LOG.info("Project present. Updating...");
-			prjToUpdate.get().setProjId(prj.getProjId());
-			prjToUpdate.get().setProjName(prj.getProjName());
-			prjToUpdate.get().setProjStatus(prj.getProjStatus());
-			LOG.info("Saving...");
-			pRepo.save(prjToUpdate.get());
-			LOG.info("Saved. Returning project");
-			return pDTO;
+		Optional<Admin> adm = aRepo.findById(adminId);
+		if (adm.isPresent()) {
+			if (prjToUpdate.isPresent()) {
+				LOG.info("Project present. Updating...");
+				prjToUpdate.get().setProjId(prj.getProjId());
+				prjToUpdate.get().setProjName(prj.getProjName());
+				prjToUpdate.get().setProjStatus(prj.getProjStatus());
+				LOG.info("Saving...");
+				pRepo.save(prjToUpdate.get());
+				LOG.info("Saved. Returning project");
+				return pDTO;
+			} else {
+				throw new NoSuchProjectFoundException(NO_PROJECT_FOUND);
+			}
 		} else {
-			throw new NoSuchProjectFoundException(NO_PROJECT_FOUND);
+			throw new NotAdminException(NOT_ADMIN);
 		}
 	}
 
 	@Override
-	public ProjectDTO deleteProject(long id) throws NoSuchProjectFoundException {
+	public ProjectDTO deleteProject(long id, long adminId) throws NoSuchProjectFoundException, NotAdminException {
 		Optional<Project> prjToDel = pRepo.findById(id);
-		if (prjToDel.isPresent())
-			pRepo.delete(prjToDel.get());
-		else
-			throw new NoSuchProjectFoundException(NO_PROJECT_FOUND);
-		return modelMapper.map(prjToDel.get(), ProjectDTO.class);
-	}
+		Optional<Admin> admDel = aRepo.findById(id);
+		if (aRepo.existsById(adminId)) {
+			if (prjToDel.isPresent()) {
+				LOG.info("Deleting...");
+				pRepo.delete(prjToDel.get());
+				LOG.info("Deleted.");
+			} else
+				throw new NoSuchProjectFoundException(NO_PROJECT_FOUND);
+			return modelMapper.map(prjToDel.get(), ProjectDTO.class);
+		} else {
+			throw new NotAdminException(NOT_ADMIN);
+		}
 
+	}
 }
